new api

+++ injection/src/plugins/transforms/double.cpp
@@ -7,7 +7,7 @@
 #include "VnV.h"
 #include "interfaces/ITransform.h"
 
-INJECTION_TRANSFORM_R(VNVPACKAGENAME, doubleToInt, double, double, int) {
+INJECTION_TRANSFORM_R(VNVPACKAGENAME, doubleToInt, int, double, int) {
   *runner = *ptr;
   return runner.get();
 }
